Refactored Workout Templates and improved History cards, close #87

+++ app/src/main/java/com/symphony/mrfit/data/model/History.kt
@@ -1,7 +1,7 @@
 /*
- *  Created by Team Symphony on 2/24/23, 11:21 PM
+ *  Created by Team Symphony on 2/25/23, 1:42 AM
  *  Copyright (c) 2023 . All rights reserved.
- *  Last modified 2/24/23, 11:20 PM
+ *  Last modified 2/25/23, 1:42 AM
  */
 
 package com.symphony.mrfit.data.model
@@ -15,5 +15,5 @@ import com.google.firebase.Timestamp
 data class History(
     val name: String = "",
     val date: Timestamp? = null,
-    val duration: Int? = null
+    val duration: Long? = null
 )
